Refactor email confirmation and password reset logic; update sender email and disable debug mode
